+++ pkg/dumper/dumper.go
@@ -14,6 +14,7 @@ import (
 	"github.com/DataDog/KubeHound/pkg/kubehound/ingestor/preflight"
 	"github.com/DataDog/KubeHound/pkg/telemetry/log"
 	"github.com/DataDog/KubeHound/pkg/telemetry/metric"
+	"github.com/DataDog/KubeHound/pkg/telemetry/span"
 	"github.com/DataDog/KubeHound/pkg/telemetry/statsd"
 	"github.com/DataDog/KubeHound/pkg/telemetry/tag"
 	"gopkg.in/DataDog/dd-trace-go.v1/ddtrace/tracer"
@@ -162,6 +163,24 @@ func dumpK8sObjs(ctx context.Context, operationName string, entity string, strea
 	err := streamFunc(ctx)
 
 	return ctx, err
+}
+
+func (d *Dumper) DumpK8sObjects(ctx context.Context) error {
+	spanDump, ctx := tracer.StartSpanFromContext(ctx, span.CollectorDump, tracer.Measured())
+	defer spanDump.Finish()
+
+	ctx, pipeline, err := newPipelineDumper(ctx, d)
+	if err != nil {
+		return fmt.Errorf("create pipeline ingestor: %w", err)
+	}
+	spanDump.SetTag(tag.DumperWorkerNumberTag, pipeline.WorkerNumber)
+
+	err = pipeline.Run(ctx)
+	if err != nil {
+		return fmt.Errorf("run pipeline ingestor: %w", err)
+	}
+
+	return pipeline.Wait(ctx)
 }
 
 func (d *Dumper) processObject(ctx context.Context, obj interface{}, filePath string) error {
